requirements.txt
appnope==0.1.3
asgiref==3.7.2
astroid==3.0.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
boto3==1.28.44
botocore==1.31.85
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
confluent-kafka==2.0.2
coverage==6.2
decorator==5.1.1
dill==0.3.7
distlib==0.3.8
Django==4.1.13
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-froala-editor==3.2.2
django-redis==5.3.0
django-unixtimestampfield==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecs-logging==2.1.0
elastic-apm==6.18.0
elasticsearch==7.13.0
executing==1.2.0
filelock==3.13.1
gunicorn==20.1.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.2
isort==5.13.2
jedi==0.18.2
jmespath==1.0.1
marshmallow==3.14.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mysqlclient==2.1.0
newrelic==8.8.0
nodeenv==1.8.0
numpy==1.23.5
packaging==23.2
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.38
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.15.1
PyJWT==2.7.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==7.4.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
responses==0.23.1
retry==0.9.2
s3transfer==0.6.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.12.3
traitlets==5.9.0
types-PyYAML==6.0.12.12
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.18
Pillow==10.2.0
virtualenv==20.25.0
wcwidth==0.2.6
wrapt==1.16.0
